+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - fixed Amazon Filter Paid Content on amazon.com (not .de)
 - fixed don't use skip intro on amazon Credits.
+- Crunchyroll: Changed timeout to include credits and intro skips.
 
 ## 1.1.86