+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.1.89
+
+- fixed Amazon Filter Paid Content
+
 ## 1.1.88
 
 - updated some dependencies